edit screens a little & make a date picker
