+++ lib/screen/registerMedInfoManual.dart
@@ -1,0 +1,8 @@
+// class RegisterMedPage
+// 복약 정보 등록하기
+import 'package:flutter/material.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter_application/constants/colors.dart';
+import 'package:flutter_application/widget/datePickerWidget.dart';
+
+// 4. 복약 정보 등록하기 화면 - 수동 등록
